Optimize pd monitor configuration write logic. (#3924)
