+++ tests/e2e/tidbcluster/tidbcluster.go
@@ -819,17 +819,6 @@ var _ = ginkgo.Describe("TiDBCluster", func() {
 		ginkgo.By("Delete tidbmonitor")
 		err = cli.PingcapV1alpha1().TidbMonitors(tm.Namespace).Delete(tm.Name, &metav1.DeleteOptions{})
 		framework.ExpectNoError(err, "delete tidbmonitor failed")
-		err = wait.Poll(5*time.Second, 3*time.Minute, func() (done bool, err error) {
-			tc, err := cli.PingcapV1alpha1().TidbClusters(tc.Namespace).Get(tc.Name, metav1.GetOptions{})
-			if err != nil {
-				return false, err
-			}
-			if tc.Status.Monitor != nil {
-				return false, nil
-			}
-			return true, nil
-		})
-		framework.ExpectNoError(err, "tc monitorRef status failed to clean after monitor deleted")
 	})
 
 	ginkgo.It("can be paused and resumed", func() {
